requirements.txt
# Speech-to-Text
openai-whisper

# Text-to-Speech
TTS

# Ollama Client
ollama

# Ses kaydı ve oynatma için gerekli olabilir
pyaudio
sounddevice

# Klavye dinleme
keyboard